+++ geo/src/algorithm/mod.rs
@@ -236,6 +236,14 @@ pub use simplify::{Simplify, SimplifyIdx};
 pub mod simplify_vw;
 pub use simplify_vw::{SimplifyVw, SimplifyVwIdx, SimplifyVwPreserve};
 
+/// Stitch together geometries with adjacent sides. Alternative to unioning geometries.
+pub mod stitch;
+pub use stitch::Stitch;
+
+/// Boolean Operations based on constrained triangulation
+pub mod spade_boolops;
+pub use spade_boolops::SpadeBoolops;
+
 /// Transform a geometry using PROJ.
 #[cfg(feature = "use-proj")]
 pub mod transform;